Add files via upload
After retrieving the number of ranges in the source DHCP, each range is exported. The name of the export xml file is constructed with the network ID of the span. After the export, the number of file and the number of ranges are compared. If the number is the same, the import is performed, otherwise a message appears and the script is exited.

The script can be executed from the source or destination.

Translated with www.DeepL.com/Translator (free version)